# public/docs/openapi.yaml
openapi: 3.0.3
info:
  title: SIGET — API de Gestión de Maquinaria y Operaciones
  description: |
    API REST para la gestión de operaciones diarias de la flota vehicular, autenticación de usuarios
    y parámetros generales del sistema.

    ## Autenticación

    La mayoría de los endpoints requieren un token **JWT**. Flujo recomendado:

    1. Llamar a `POST /auth/login` con `email` y `password` para obtener un `access_token`.
    2. Enviar dicho token en la cabecera `Authorization: Bearer {access_token}` en cada petición protegida.
    3. Cuando el token esté por expirar, llamar a `POST /auth/refresh` (con el token actual todavía vigente)
       para obtener uno nuevo.
    4. Llamar a `POST /auth/logout` para invalidar el token de forma explícita.

    ## Formato de errores

    - **422 Unprocessable Content**: error de validación. Responde `{ "success": false, "message": "...", "errors": { "campo": ["mensaje"] } }`.
    - **401 Unauthorized**: token ausente, inválido o expirado.
    - **403 Forbidden**: el usuario está autenticado pero no tiene permiso sobre el recurso solicitado.
    - **404 Not Found**: el recurso indicado en la ruta no existe.
    - **500 Internal Server Error**: error inesperado del servidor. Responde `{ "message": "...", "error": "..." }`.

    ## Usuarios de prueba

    Estas cuentas están precargadas en el entorno de pruebas y sirven para probar la
    API con cada rol desde `POST /auth/login`:

...
  version: "1.9.0"
  contact:
...
  license:
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /cargas:
    get:
      tags: [Carga de Combustible]
      summary: Listar cargas de combustible
      description: |
        Lista paginada de cargas de combustible, ordenadas por `fecha_carga` e `id` descendente.

        Las restricciones de visibilidad dependen del rol del usuario autenticado:
        - **conductor**: solo ve sus propias cargas (`id_conductor` = persona del usuario).
        - Otros roles: sin restricción adicional.
      parameters:
        - name: nro_placa
          in: query
          description: Filtra por placa o código del vehículo (coincidencia parcial).
          schema: { type: string }
          example: 1234-ABC
        - name: fecha_desde
          in: query
          description: Filtra cargas cuya `fecha_carga` sea igual o posterior a esta fecha.
          schema: { type: string, format: date }
          example: "2026-08-01"
        - name: fecha_hasta
          in: query
          description: Filtra cargas cuya `fecha_carga` sea igual o anterior a esta fecha.
          schema: { type: string, format: date }
          example: "2026-08-31"
        - name: tipo_carga
          in: query
          description: Filtra por tipo exacto de carga.
          schema: { type: string, enum: [VALE, PREPAGO] }
          example: PREPAGO
        - name: estado_carga
          in: query
          description: Filtra por estado exacto de la carga.
          schema: { type: string, enum: [REGISTRADO, VERIFICADO, ANULADO] }
          example: REGISTRADO
        - name: per_page
          in: query
          description: Cantidad de resultados por página.
          schema: { type: integer, default: 10 }
          example: 10
        - name: page
          in: query
          description: |
            Número de página a obtener (paginación estándar del listado, se resuelve
            automáticamente a partir de este parámetro). Si se omite, se asume `1`.
          schema: { type: integer, default: 1, minimum: 1 }
          example: 1
      responses:
        '200':
          description: Listado paginado de cargas de combustible.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CargaCombustiblePaginada'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags: [Carga de Combustible]
      summary: Registrar una nueva carga de combustible
      description: |
        Registra una carga de combustible junto con sus respaldos digitales.

        **Debe enviarse obligatoriamente como `multipart/form-data`** (no `application/json`),
        ya que el campo `respaldos` adjunta uno o más archivos (imagen o PDF) por carga.

        `id_usuario` se asigna automáticamente al usuario autenticado en el servidor y no debe
        enviarse. Lo mismo ocurre con `nro_carga` y `gestion`: correlativo autogenerado dentro
        de la gestión actual (configurable en los parámetros de la empresa; por defecto la
        gestión avanza en diciembre), expuesto en la respuesta de `GET /cargas` como `nro`
        (`nro_carga` formateado con ceros a la izquierda seguido de `/gestion`).

        ## Reglas condicionales relevantes

        - `litros` y `precio` son obligatorios **solo si no se envía `id_vale`**. Si se envía
          `id_vale`, ambos valores se ignoran y se copian automáticamente desde el vale
          referenciado.
        - `kilometraje` es obligatorio si el vehículo indicado (`id_vehiculo`) tiene
          `tipo_medicion = kilometraje`; `horometro` es obligatorio si tiene
          `tipo_medicion = horometro`. En ambos casos, el valor no puede ser inferior a la
          última lectura registrada para ese vehículo.
        - `id_conductor` es obligatorio si el usuario autenticado tiene el rol `conductor`;
          opcional para el resto de roles.
        - Si se envía `id_vale`, debe referenciar un vale con `estado_vale = PENDIENTE`, del
          mismo `id_vehiculo`, y aún no vencido (`fecha_vencimiento > ahora`). Al registrarse la
          carga, ese vale queda marcado como `USADO`.
        - `concepto` ("A utilizarse en") es un texto libre y opcional que describe para qué se
          usará el combustible cargado (ej. "Traslado de personal a planta norte").
        - `is_offline` (uso exclusivo de la app Android, que permite registrar cargas sin
          conexión y sincronizarlas después), si se envía `true`:
          - Omite, solo para esta petición, la exigencia anterior de que `id_vale` siga
            `PENDIENTE` y vigente, ya que para el momento de la sincronización el vale pudo
            haber sido usado o vencido.
          - Hace que `fecha_carga` pase a ser **obligatoria**, ya que en ese caso corresponde
            a la fecha real en que se generó la carga offline (antes de sincronizarse), y no a
            la fecha de sincronización. Si se omite o se envía `false`, `fecha_carga` es
            opcional.

        ## Envío de los respaldos digitales (`respaldos`)

        `respaldos` es un arreglo; cada elemento agrega **dos** campos al `multipart/form-data`,
        con `archivo` (el binario) y `tipo` (texto):

        ```
        respaldos[0][archivo] = (factura.jpg)
        respaldos[0][tipo]    = FACTURA
        respaldos[1][archivo] = (nota.pdf)
        respaldos[1][tipo]    = NOTA
        ```

        Por cada elemento con archivo presente: el tipo de archivo (`IMAGEN` o `PDF`) se
        detecta automáticamente según el archivo recibido, y se guarda asociado con
        `tipo_respaldo` = el `tipo` enviado para ese mismo índice (`OTRO` si se omite). Ver los
        valores válidos de `tipo` en `GET /parametros/colecciones` →
        `cargas_combustible.tipos_respaldo_digital`.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/CargaCombustibleStoreRequest'
      responses:
        '201':
          description: Carga de combustible creada exitosamente.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CargaCombustibleStoreResponse'
              example:
                message: Carga de combustible creada exitosamente
        '422':
          description: Error de validación.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DefaultValidationErrorResponse'
              example:
                message: The given data was invalid.
                errors:
                  fecha_carga: ["El campo fecha de carga es obligatorio."]
                  id_vehiculo: ["El campo vehículo es obligatorio."]
                  tipo_carga: ["El campo tipo de carga es obligatorio."]
                  kilometraje: ["El kilometraje es obligatorio para vehículos con medición por kilometraje."]
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          description: Error inesperado al crear la carga de combustible.
          content:
            application/json:
              example:
                message: Error al crear la carga de combustible.
                error: "Detalle técnico interno del error (varía según la causa)."
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        Token JWT obtenido en `POST /auth/login`. Enviar como
        `Authorization: Bearer {access_token}`.
  responses:
    Unauthorized:
      description: No autenticado, token ausente/inválido/expirado.
      content:
        application/json:
          example:
            message: Unauthenticated.
...
  schemas:
...
    DefaultValidationErrorResponse:
      type: object
      description: |
        Formato **estándar** de error de validación (sin personalizar). A diferencia de
        `ValidationErrorResponse` (usado por algunos endpoints, como el de operación diaria),
        aquí **no** se envuelve con la clave `success`, y el `message` de nivel superior queda
        fijo en inglés (`"The given data was invalid."`) sin traducir. Los mensajes dentro de
        `errors` sí están en español.
      properties:
        message: { type: string, example: "The given data was invalid." }
        errors:
          type: object
          additionalProperties:
            type: array
            items: { type: string }
          example:
            campo: ["mensaje de error"]

    # ── Carga de Combustible ─────────────────────────────────────────────
    GrifoCompleto:
      type: object
      description: Grifo/estación de servicio completa, a diferencia de `Grifo` (usado en `colecciones`, con columnas renombradas y restringidas).
      properties:
        id: { type: integer, example: 2 }
        razon_social: { type: string, example: Gasolinera Central S.A. }
        nit: { type: string, example: "1234567890" }
        direccion: { type: string, nullable: true, example: "Av. 6 de Agosto" }
        ciudad: { type: string, nullable: true, example: La Paz }
        telefono: { type: string, nullable: true, example: "22345678" }
        estado_grifo: { type: string, enum: [ACTIVO, INACTIVO], example: ACTIVO }
        es_principal: { type: boolean, example: false }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    TipoCombustibleCompleto:
      type: object
      description: Tipo de combustible completo, a diferencia de `TipoCombustible` (usado en `colecciones`, con columnas renombradas).
      properties:
        id: { type: integer, example: 3 }
        tipo_combustible: { type: string, example: Diésel }
        estado_tipo_combustible: { type: string, enum: [ACTIVO, INACTIVO], example: ACTIVO }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
    CargaCombustible:
      type: object
      properties:
        id: { type: integer, example: 15 }
        nro_carga: { type: integer, description: Correlativo autogenerado dentro de la gestión., example: 1 }
        gestion: { type: string, example: "2026" }
        nro: { type: string, description: "`nro_carga` formateado con ceros a la izquierda (cantidad de dígitos configurable en los parámetros de la empresa) seguido de `/gestion`.", example: "000001/2026" }
        fecha_carga: { type: string, description: "Formateada como `Y-m-d H:i` por el cast del modelo.", example: "2026-08-17 14:30" }
        litros: { type: string, example: "50.00" }
        precio: { type: string, example: "10.00" }
        kilometraje: { type: string, nullable: true, example: "15600.00" }
        horometro: { type: string, nullable: true, example: null }
        id_vehiculo: { type: integer, example: 8 }
        id_grifo: { type: integer, example: 2 }
        id_tipo_combustible: { type: integer, example: 3 }
        id_conductor: { type: integer, example: 1 }
        id_vale: { type: integer, nullable: true, example: null }
        nro_factura: { type: string, nullable: true, example: "F-00123" }
        concepto: { type: string, nullable: true, description: "\"A utilizarse en\": para qué se usará el combustible cargado.", example: "Traslado de personal a planta norte" }
        tipo_carga: { type: string, enum: [VALE, PREPAGO], example: PREPAGO }
        estado_carga: { type: string, nullable: true, enum: [REGISTRADO, VERIFICADO, ANULADO], example: REGISTRADO }
        id_usuario: { type: integer, nullable: true, description: Usuario que registró la carga (asignado automáticamente al crear)., example: 2 }
        fecha_carga_formateada: { type: string, description: "`fecha_carga` formateada como `d/m/Y H:i`.", example: "17/08/2026 14:30" }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        vehiculo: { $ref: '#/components/schemas/Vehiculo' }
        grifo: { $ref: '#/components/schemas/GrifoCompleto' }
        tipoCombustible: { $ref: '#/components/schemas/TipoCombustibleCompleto' }
        conductor: { $ref: '#/components/schemas/Conductor' }
        vale:
          description: |
            Presente solo si la carga se generó a partir de un vale (`id_vale` no nulo). A
            diferencia del recurso `Vale` de `GET /vales`, aquí no vienen cargadas las
            relaciones anidadas del vale (`vehiculo`, `conductor`, `grifo`).
          nullable: true
          allOf:
            - $ref: '#/components/schemas/Vale'
    CargaCombustiblePaginada:
      type: object
      description: Estructura estándar de paginación.
      properties:
        current_page: { type: integer, example: 1 }
        data:
          type: array
          items: { $ref: '#/components/schemas/CargaCombustible' }
        first_page_url: { type: string, example: "http://localhost/api/v1/cargas?page=1" }
        from: { type: integer, nullable: true, example: 1 }
        last_page: { type: integer, example: 1 }
        last_page_url: { type: string, example: "http://localhost/api/v1/cargas?page=1" }
        links:
          type: array
          items:
            type: object
            properties:
              url: { type: string, nullable: true }
              label: { type: string, example: "1" }
              active: { type: boolean, example: true }
        next_page_url: { type: string, nullable: true, example: null }
        path: { type: string, example: "http://localhost/api/v1/cargas" }
        per_page: { type: integer, example: 10 }
        prev_page_url: { type: string, nullable: true, example: null }
        to: { type: integer, nullable: true, example: 1 }
        total: { type: integer, example: 1 }
    CargaCombustibleStoreRequest:
      type: object
      description: |
        Cuerpo esperado por `POST /cargas`, enviado como `multipart/form-data`. `id_usuario`
        se asigna automáticamente al usuario autenticado y no debe enviarse.
      required: [id_vehiculo, id_grifo, id_tipo_combustible, tipo_carga]
      properties:
        fecha_carga:
          type: string
          format: date-time
          nullable: true
          description: |
            Opcional si `is_offline` se omite o es `false`. **Obligatoria si `is_offline =
            true`** (fecha real en que se generó la carga offline, antes de sincronizarse).
          example: "2026-08-17 14:30"
        litros:
          type: number
          format: float
          nullable: true
          minimum: 0.01
          maximum: 9999.99
          description: Obligatorio si no se envía `id_vale`. Si se envía `id_vale`, este valor se ignora y se toma automáticamente del vale.
          example: 50
        precio:
          type: number
          format: float
          nullable: true
          minimum: 0.01
          maximum: 99999.99
          description: Obligatorio si no se envía `id_vale`. Si se envía `id_vale`, este valor se ignora y se toma automáticamente del vale.
          example: 10
        kilometraje:
          type: number
          format: float
          nullable: true
          minimum: 0
          description: Obligatorio si el vehículo (`id_vehiculo`) tiene `tipo_medicion = kilometraje`. No puede ser menor a la última lectura registrada para ese vehículo.
          example: 15600
        horometro:
          type: number
          format: float
          nullable: true
          minimum: 0
          description: Obligatorio si el vehículo (`id_vehiculo`) tiene `tipo_medicion = horometro`. No puede ser menor a la última lectura registrada para ese vehículo.
          example: null
        id_vehiculo:
          type: integer
          description: Debe existir en la tabla `vehiculo`.
          example: 8
        id_grifo:
          type: integer
          description: requerido en caso de que no se envíe `id_vale`. Debe existir en la tabla `grifo` y tener `estado_grifo = ACTIVO`.
          example: 2
        id_tipo_combustible:
          type: integer
          description: requerido en caso de que no se envíe `id_vale`. Debe existir en la tabla `tipo_combustible` y tener `estado_tipo_combustible = ACTIVO`.
          example: 3
        id_vale:
          type: integer
          nullable: true
          description: |
            Si se envía, debe referenciar un vale con `estado_vale = PENDIENTE`, del mismo
            `id_vehiculo`, y aún no vencido. Al registrar la carga, el vale queda marcado como
            `USADO` y sus `litros`/`precio` se copian a la carga.
          example: null
        nro_factura:
          type: string
          nullable: true
          maxLength: 50
          example: "F-00123"
        concepto:
          type: string
          nullable: true
          maxLength: 255
          description: "\"A utilizarse en\": describe para qué se usará el combustible cargado."
          example: "Traslado de personal a planta norte"
        tipo_carga:
          type: string
          enum: [VALE, PREPAGO]
          description: Tipo de carga de combustible. Si se envía `VALE`, `id_vale` es obligatorio y `litros`/`precio` se ignoran. Si se envía `PREPAGO`, `id_vale` debe ser nulo y `litros`/`precio` son obligatorios.
          example: PREPAGO
        estado_carga:
          type: string
          nullable: true
          enum: [REGISTRADO, VERIFICADO, ANULADO]
          description: Opcional; si se omite, la carga se crea con estado `REGISTRADO`.
          example: REGISTRADO
        is_offline:
          type: boolean
          nullable: true
          default: false
          description: |
            Uso exclusivo de la app Android (Flutter), que permite registrar cargas de
            combustible sin conexión y sincronizarlas después. Si se envía `true`:
            - Se omite la exigencia de que `id_vale` siga con `estado_vale = PENDIENTE` y
              `fecha_vencimiento` vigente, ya que para el momento en que se sincroniza el vale
              pudo haber sido usado o vencido.
            - `fecha_carga` pasa a ser obligatoria (ver descripción de ese campo).

            No tiene efecto en peticiones fuera de la API (p. ej. el panel web) ni al editar
            una carga sin cambiar su `id_vale`.
          example: false
        respaldos[0][archivo]:
          type: string
          format: binary
          description: |
            Archivo (imagen o PDF) del primer respaldo digital. Repetir incrementando el
            índice (`respaldos[1][archivo]`, `respaldos[2][archivo]`, ...) para adjuntar más de
            un respaldo. El tipo de archivo (`IMAGEN`/`PDF`) se detecta automáticamente por su
            MIME type. hasta 5 archivos. Tamaño máximo por archivo: 2 MB.
        respaldos[0][tipo]:
          type: string
          enum: [FACTURA, NOTA, COMPROBANTE, OTRO]
          default: OTRO
          description: Tipo de respaldo correspondiente a cada `respaldos[0][archivo]`.
          example: FACTURA
    CargaCombustibleStoreResponse:
      type: object
      description: A diferencia de otros endpoints de registro de esta API, aquí la respuesta **no** incluye la carga creada (`data`).
      properties:
        message: { type: string, example: "Carga de combustible creada exitosamente" }

    # ── Control de Cargas ────────────────────────────────────────────────
...